increase max tokens limit
increases the maximum amount of tokens allowed for the `text-davinci-003` model from 256 to 2000. this will allow for larger tasks to be completed in the same amount of time.

this commit makes the following changes:

- updates `max_tokens` in `defaultconfigyml` from `256` to `2000`

the change is necessary in order to allow for larger tasks to be completed in the same amount of time. increasing the `max_tokens` limit from `256` to `2000` will allow for a more efficient workflow.

this change will have a positive impact on the performance of the `text-davinci-003` model. by increasing the `max_tokens` limit, users will be able to complete more tasks in the same amount of time.