update readme.md
this commit updates the readme.md file to reflect the changes in the cligpt repository. the changes include updating the installation instructions to use the latest version of the repository, updating the example command to use bash syntax, and removing an image.

```diff
diff --git a/readme.md b/readme.md
index 3134605cc697..d1684524afe7 100644
--- a/readme.md
+++ b/readme.md
@@ -4,14 +4,13 @@

 go 1.17+
 ```
-$ go install github.com/paij0se/cligpt@latest
+go install github.com/falcucci/cligpt@latest
 ```

 test

-```
-❯ cligpt "do a poem about golang in latin"
-+```bash
+cligpt "do a poem about golang in latin"

 nunc est tempus, ut in golang gradimur
 mox nostris verbis linguam didicimus
@@ -43,12 +42,10 @@ max_tokens: 2000

  - clone the repository.

-`$ git clone https://github.com/paij0se/cligpt`
+`git clone https://github.com/falcucci/cligpt`

  - build the cli.

 ```bash
 go build
 ```
-
-![image](https://user-images.githubusercontent.com/69026987/209194859-a2456a7d-796f-47e0-8a8e-062848e2cbaf.png)
```